requirements.txt
# Django
Django==5.1.4
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-widget-tweaks==1.5.0

# Django REST Framework
djangorestframework==3.15.2

# Database
psycopg[binary]==3.2.3
pgvector==0.3.6

# Celery
celery[redis]==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Storage (S3/MinIO)
boto3==1.35.71
django-storages==1.14.4

# AI / LLM
google-generativeai==0.8.5
tiktoken==0.8.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0

# HTTP / API clients
httpx==0.28.1
tenacity==9.0.0

# Utilities
python-dateutil==2.9.0
hashlib-additional==1.0.1

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.1
responses==0.25.3

# Code quality
ruff==0.8.2

# WSGI
gunicorn==23.0.0
whitenoise==6.8.2
